Move list helpers to separate module
