+++ src/app/features/editor/selectors.ts
@@ -1,19 +1,22 @@
 import { createSelector } from "@reduxjs/toolkit";
-import { omit } from "lodash";
 import type {
-  EventId,
   CardId,
   DeckId,
   EntityId,
-  PropertyId,
+  EventId,
   MiddlewareId,
+  PropertyId,
 } from "../../../api/services/game/types";
 import { getKeyVisibilities } from "../../../lib/reactMosaicExtensions";
 import { compileEditorApi } from "../compiler/compileEditorApi";
 import { deriveRuntimeDefinition } from "../compiler/defineRuntime";
 import { gameTypes } from "../gameTypes";
-import { adjacent } from "../../../lib/adjacent";
 import type { EditorObjectId, EditorState, EditorSyncState } from "./types";
+import {
+  adjacentSelectedObject,
+  objectById,
+  selectedList,
+} from "./utils/lists";
 
 const gameDefinition = (state: EditorState) => state.game?.definition;
 
@@ -30,53 +33,6 @@ const runtimeDefinition = createSelector(
 const editorApi = createSelector(runtimeDefinition, (def) =>
   def ? compileEditorApi(def) : undefined
 );
-
-function adjacentSelectedObject(
-  state: EditorState
-): EditorObjectId | undefined {
-  const id = state.selectedObjectId;
-  const list = selectedList(state);
-  const obj = objectById(state.selectedObjectId)(state);
-  const adj = adjacent(list, obj);
-  if (id && adj) {
-    const prop = getObjectIdProperty(id);
-    return { ...id, [prop]: adj[prop] };
-  }
-}
-
-function selectedList(state: EditorState) {
-  switch (state.selectedObjectId?.type) {
-    case "event":
-      return state.game?.definition.events;
-    case "middleware":
-      return state.game?.definition.middlewares;
-    case "card":
-      return state.game?.definition.cards;
-    case "deck":
-      return state.game?.definition.decks;
-  }
-}
-
-function objectById(id?: EditorObjectId) {
-  return (state: EditorState) => {
-    switch (id?.type) {
-      case "event":
-        return selectors.event(id.eventId)(state);
-      case "middleware":
-        return selectors.middleware(id.middlewareId)(state);
-      case "card":
-        return selectors.card(id.cardId)(state);
-      case "deck":
-        return selectors.deck(id.deckId)(state);
-    }
-  };
-}
-
-function getObjectIdProperty<T extends EditorObjectId>(id: T) {
-  type IdProperty = keyof Omit<T, "type">;
-  const idProperty = Object.keys(omit(id, "type"))[0];
-  return idProperty as IdProperty;
-}
 
 export const selectors = {
   isSyncState:
